+++ web/packages/new/photos/components/gallery/reducer.ts
@@ -28,6 +28,7 @@ import type {
     CollectionSummary,
     CollectionSummaryType,
 } from "../../services/collection/ui";
+import type { FamilyData } from "../../services/family";
 import {
     createFileCollectionIDs,
     getLatestVersionFiles,
@@ -40,7 +41,6 @@ import {
     isPinnedCollection,
 } from "../../services/magic-metadata";
 import type { PeopleState, Person } from "../../services/ml/people";
-import type { FamilyData } from "../../services/user";
 
 /**
  * Specifies what the bar at the top of the gallery is displaying currently.
