+++ web/packages/new/photos/services/user.ts
@@ -1,13 +1,7 @@
 import { authenticatedRequestHeaders, ensureOk } from "@/base/http";
-import log from "@/base/log";
 import { apiURL } from "@/base/origins";
-import {
-    LS_KEYS,
-    getData,
-    removeData,
-} from "@ente/shared/storage/localStorage";
-import type { User } from "@ente/shared/user/types";
 import { z } from "zod";
+import type { FamilyData } from "./family";
 import type { Subscription } from "./plan";
 
 export interface Bonus {
@@ -31,75 +25,6 @@ export interface UserDetails {
     storageBonus?: number;
     bonusData?: BonusData;
 }
-
-export interface FamilyMember {
-    email: string;
-    usage: number;
-    id: string;
-    isAdmin: boolean;
-}
-
-export interface FamilyData {
-    storage: number;
-    expiry: number;
-    members: FamilyMember[];
-}
-
-export function getLocalFamilyData(): FamilyData {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-return
-    return getData(LS_KEYS.FAMILY_DATA);
-}
-
-// isPartOfFamily return true if the current user is part of some family plan
-export function isPartOfFamily(familyData: FamilyData): boolean {
-    return Boolean(
-        // eslint-disable-next-line @typescript-eslint/prefer-optional-chain, @typescript-eslint/no-unnecessary-condition
-        familyData && familyData.members && familyData.members.length > 0,
-    );
-}
-
-// hasNonAdminFamilyMembers return true if the admin user has members in his family
-export function hasNonAdminFamilyMembers(familyData: FamilyData): boolean {
-    return Boolean(isPartOfFamily(familyData) && familyData.members.length > 1);
-}
-
-export function isFamilyAdmin(familyData: FamilyData): boolean {
-    const familyAdmin: FamilyMember = getFamilyPlanAdmin(familyData);
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const user: User = getData(LS_KEYS.USER);
-    return familyAdmin.email === user.email;
-}
-
-export function getFamilyPlanAdmin(familyData: FamilyData): FamilyMember {
-    if (isPartOfFamily(familyData)) {
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        return familyData.members.find((x) => x.isAdmin)!;
-    } else {
-        log.error(
-            "invalid getFamilyPlanAdmin call - verify user is part of family plan before calling this method",
-        );
-        throw new Error(
-            "invalid getFamilyPlanAdmin call - verify user is part of family plan before calling this method",
-        );
-    }
-}
-
-export function getTotalFamilyUsage(familyData: FamilyData): number {
-    return familyData.members.reduce(
-        (sum, currentMember) => sum + currentMember.usage,
-        0,
-    );
-}
-
-export const leaveFamily = async () => {
-    ensureOk(
-        await fetch(await apiURL("/family/leave"), {
-            method: "DELETE",
-            headers: await authenticatedRequestHeaders(),
-        }),
-    );
-    removeData(LS_KEYS.FAMILY_DATA);
-};
 
 /**
  * Fetch the two-factor status (whether or not it is enabled) from remote.
